Fix a label update issue reported by a customer.
Summary:
This one is pretty fun ;) One customer had a problem was that marking a message as read would work but moving it to the trash wouldn't. Here's what I think happened:

1. The customer marks a thread as read and moves it to the trash. It works.
2. he receives a new email related to this thread. Gmail's behaviour in this case is to move the thread back into Inbox and to discard its "trash" label. We don't pick this up.
3. He repeats 1. This time moving to trash doesn't work because we think the thread is already in the trash.

My solution was to manually discard the trash folder from a thread's folders if the newest UID doesn't have it. I understand that the code isn't super pretty and I'm open to alternatives.

Test Plan: Tested manually.

Reviewers: spang, kav-ya, emfree

Reviewed By: emfree

Projects: #sync-engine

Differential Revision: https://review.inboxapp.com/D1121