[gmail] Correctly update read status for already downloaded ImapUids
Summary:
We don't update the read status when creating an ImapUid for a message we already have. This is a problem in the following scenario:
Let's say I have a new, unread message in my Inbox. I read it quickly and move it to the trash. We don't pick up that the message's read status has changed.

This diff fixes this.

Test Plan: Tested manually.

Reviewers: spang, jennie, emfree, kav-ya

Reviewed By: kav-ya

Projects: #sync-engine

Differential Revision: https://review.inboxapp.com/D1244