Make our threading algorithm a little more smarter
Summary:
This is a potential fix for the email threading issue a customer reported this morning. I'm saying "potential" because it's going to be more expensive in term of requests than what we already do.

Previously we would blindly group threads together if they had the same subject. This diff changes this to the following:
1. Look if the message we're "threading" is a reply to an existing message
2. Look if there's a reply to the message we're threading
3. Look if there's thread who share the same title and a subset of participants with the current message.

I think this should be enough for most cases.

Test Plan: Tested manually on my gandi account. Checked that threads were weaved correctly.

Reviewers: kav-ya, spang, emfree

Reviewed By: emfree

Subscribers: mg

Projects: #customers, #sync-engine

Differential Revision: https://review.inboxapp.com/D1243

Conflicts:
	bin/compute_cleaned_subject
	inbox/mailsync/backends/imap/generic.py
	inbox/models/message.py
	inbox/util/threading.py
	tests/data/base_dump.sql
	tests/general/test_threading.py