Issue first heartbeat outside retry loop in IMAP sync.
Summary:
Otherwise, if there's a problem starting the sync, the folder sync may
repeatedly "eagerly signal" its status as it retries, meaning that it appears
alive when it is in fact dead.

Test Plan:
Break sync by raising exception in ImapFolderSyncEngine._run_impl;
check that it stays dead in admin dashboard.

Reviewers: jennie

Reviewed By: jennie

Differential Revision: https://review.inboxapp.com/D1304